Add some code comments and cleanup.
