+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/datamodel/node/InstanceSet.scala
@@ -9,8 +9,6 @@ package edu.illinois.cs.cogcomp.saul.datamodel.node
 import edu.illinois.cs.cogcomp.saul.datamodel.edge.Edge
 import edu.illinois.cs.cogcomp.saul.datamodel.property.TypedProperty
 
-import scala.collection.mutable.ArrayBuffer
-
 trait InstanceSet[T <: AnyRef] extends Iterable[T] {
   self =>
   def instances: Iterable[T]
@@ -53,19 +51,4 @@ case class NodeSet[T <: AnyRef](node: Node[T]) extends InstanceSet[T] {
 
 case class SingletonSet[T <: AnyRef](node: Node[T], t: T) extends InstanceSet[T] {
   val instances: Iterable[T] = Set(t)
-}
-
-trait PropertySet[T <: AnyRef, V] extends Iterable[V] {
-  self =>
-  def property: TypedProperty[T, V]
-  def underlying: InstanceSet[T]
-  lazy val propValues: Iterable[V] = {
-    val ab = new ArrayBuffer[V]
-    ab ++= underlying.map(property(_))
-    ab
-  }
-
-  override def iterator: Iterator[V] = propValues.iterator
-
-  def counts = propValues.groupBy(x => x).map(p => p._1 -> p._2.size).toMap
 }
