+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/datamodel/property/features/real/RealArrayProperty.scala
@@ -12,10 +12,14 @@ import edu.illinois.cs.cogcomp.saul.datamodel.property.features.ClassifierContai
 
 import scala.reflect.ClassTag
 
-case class RealArrayProperty[T <: AnyRef](
-  name: String,
-  sensor: T => List[Double]
-)(implicit val tag: ClassTag[T]) extends RealPropertyCollection[T] with TypedProperty[T, List[Double]] {
+/** Represents Real Valued Array attributes.
+  *
+  * @param name Name of the property
+  * @param sensor Sensor function used to generate attributes from nodes.
+  * @param tag ClassTag for the type of data stored by the attribute node
+  * @tparam T
+  */
+case class RealArrayProperty[T <: AnyRef](name: String, sensor: T => List[Double])(implicit val tag: ClassTag[T]) extends RealPropertyCollection[T] with TypedProperty[T, List[Double]] {
 
   // TODO: shouldn't this be this.name?
   val ra = this
@@ -38,7 +42,7 @@ case class RealArrayProperty[T <: AnyRef](
       val d: T = example.asInstanceOf[T]
       val values = sensor(d)
 
-      var featureVector = new FeatureVector
+      val featureVector = new FeatureVector
 
       values.zipWithIndex.foreach {
         case (value, idx) => featureVector.addFeature(new RealArrayStringFeature(this.containingPackage, this.name, "", value, idx, 0))
