+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/lbjrelated/LBJClassifierEquivalent.scala
@@ -8,8 +8,8 @@ package edu.illinois.cs.cogcomp.saul.lbjrelated
 
 import edu.illinois.cs.cogcomp.lbjava.classify.Classifier
 
-import edu.illinois.cs.cogcomp.saul.constraint.LHSFirstOrderEqualityWithValueLBP
-
+/** Encapsulates an instance of LBJava's [[Classifier]] class.
+  */
 trait LBJClassifierEquivalent {
   val classifier: Classifier
 }
