+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/lbjrelated/LBJLearnerEquivalent.scala
@@ -10,10 +10,17 @@ import edu.illinois.cs.cogcomp.lbjava.classify.Classifier
 import edu.illinois.cs.cogcomp.lbjava.learn.{ Lexicon, Learner }
 import edu.illinois.cs.cogcomp.saul.constraint.LHSFirstOrderEqualityWithValueLBP
 
+/** Encapsulates an instance of LBJava's [[Learner]] class.
+  */
 trait LBJLearnerEquivalent extends LBJClassifierEquivalent {
+
   val classifier: Learner
+
   def on(t: AnyRef): LHSFirstOrderEqualityWithValueLBP = new LHSFirstOrderEqualityWithValueLBP(this.classifier, t)
+
   def getLabeler: Classifier = classifier.getLabeler
+
   def getExampleArray(example: Any): Array[AnyRef] = classifier.getExampleArray(example)
+
   def getLabelLexicon: Lexicon = classifier.getLabelLexicon
 }
