+++ saul-core/src/main/scala/edu/illinois/cs/cogcomp/saul/parser/LBJavaParserToIterable.scala
@@ -11,15 +11,20 @@ import edu.illinois.cs.cogcomp.lbjava.parse.{ FoldSeparator, Parser }
 /** a utility class to convert LBJava parser to Scala iterator */
 class LBJavaParserToIterable[T <: AnyRef](parser: Parser) extends Iterable[T] {
   override def iterator(): Iterator[T] = new parserToIterator(parser)
+
   private class parserToIterator(parser: Parser) extends Iterator[T] {
     val nextBuffer = new collection.mutable.Stack[AnyRef]
+
     override def hasNext: Boolean = {
       val nextElement = parser.next()
-      if (nextElement == null || nextElement == FoldSeparator.separator) false else {
+      if (nextElement == null || nextElement == FoldSeparator.separator) {
+        false
+      } else {
         nextBuffer.push(nextElement)
         true
       }
     }
+
     override def next(): T = nextBuffer.pop().asInstanceOf[T]
   }
 }
